Trying to find the right place for functions, getting rid of import loops.
