+++ setup-MoinMoin.py
@@ -265,6 +265,8 @@ Topic :: Text Processing :: Markup""".splitlines(),
         'MoinMoin.logfile',
         'MoinMoin.macro',
         'MoinMoin.mail',
+        'MoinMoin.metadata',
+        'MoinMoin.metadata.backend',
         'MoinMoin.parser',
         'MoinMoin.script',
         'MoinMoin.script.account',
